Fix a regression.
If the user set a custom title bar height, just return
the original value, don't plus the border height afterwards,
this is not what the user would want.

But if we are querying the standard value, it should be
border height + caption height.

Border width and border height don't have this issue.

Signed-off-by: Yuhang Zhao <2546789017@qq.com>